refactor: rename MalwareScanScope to ScanScope and extract config utils
