+++ example/package.json
@@ -47,7 +47,7 @@
     "cordova-plugin-ionic-keyboard": "^2.0.5",
     "cordova-plugin-splashscreen": "5.0.2",
     "cordova-plugin-statusbar": "^2.4.2",
-    "cordova-talsec-plugin-freerasp": "file:temp_plugin/package",
+    "cordova-talsec-plugin-freerasp": "file:..",
     "eslint": "^8.57.0",
     "eslint-plugin-import": "^2.29.1",
     "eslint-plugin-jsdoc": "^48.2.1",
