+++ www/src/types/types.ts
@@ -22,7 +22,7 @@ export type ScopeType =
 
 export type ReasonMode = 'ALL' | 'HIGHEST_CONFIDENCE';
 
-export type MalwareScanScope = {
+export type ScanScope = {
   scanScope: ScopeType;
   trustedInstallSources?: string[];
 };
@@ -32,7 +32,7 @@ export type SuspiciousAppDetectionConfig = {
   hashes?: string[];
   requestedPermissions?: string[][];
   grantedPermissions?: string[][];
-  malwareScanScope: MalwareScanScope;
+  scanScope: ScanScope;
   reasonMode: ReasonMode;
 };
