+++ www/src/utils/config.ts
@@ -1,0 +1,40 @@
+import {
+  ReasonMode,
+  ScanScope,
+  SuspiciousAppDetectionConfig,
+  TalsecAndroidConfig,
+  TalsecConfig,
+} from '../types/types';
+
+const DEFAULT_SCAN_SCOPE: ScanScope = {
+  scanScope: 'SIDELOADED_ONLY',
+};
+const DEFAULT_REASON_MODE: ReasonMode = 'HIGHEST_CONFIDENCE';
+
+const withDefaults = (
+  config: SuspiciousAppDetectionConfig,
+): SuspiciousAppDetectionConfig => ({
+  ...config,
+  scanScope: config.scanScope ?? DEFAULT_SCAN_SCOPE,
+  reasonMode: config.reasonMode ?? DEFAULT_REASON_MODE,
+});
+
+const normalizeAndroidConfig = (
+  androidConfig: TalsecAndroidConfig,
+): TalsecAndroidConfig => {
+  if (!androidConfig.suspiciousAppDetectionConfig) return androidConfig;
+  return {
+    ...androidConfig,
+    suspiciousAppDetectionConfig: withDefaults(
+      androidConfig.suspiciousAppDetectionConfig,
+    ),
+  };
+};
+
+export const normalizeConfig = (config: TalsecConfig): TalsecConfig => {
+  if (!config.androidConfig) return config;
+  return {
+    ...config,
+    androidConfig: normalizeAndroidConfig(config.androidConfig),
+  };
+};
